ml-service/requirements.txt
flask
numpy
pandas
scikit-learn
joblib
tensorflow-cpu
